pyproject.toml
[project]
name = "eisenhour-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "sqlalchemy>=2.0.41",
    "pytest-cov>=6.1.1",
    "uvicorn>=0.34.2",
    "authlib>=1.6.0",
    "pyjwt>=2.10.1",
]


[tool.uv]
package = true


[tool.ruff.lint]
extend-select = ["UP", "I", "N", "RUF", "TRY"]

[tool.mypy]
python_version = "3.13"
mypy_path = "src"
strict = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true
explicit_package_bases = true

[tool.pytest.ini_options]
addopts = "--import-mode=importlib"
python_files = [
    "test_*.py"
]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit",
    "integration",
    "e2e"
]

[tool.coverage.run]
relative_files = true

[tool.coverage.paths]
source = [
    "src/",  
    "app/src"
]
